+++ packages/token-tools/CHANGELOG.md
@@ -69,6 +69,8 @@
 
 - [#589](https://github.com/terrazzoapp/terrazzo/pull/589) [`8f32d44`](https://github.com/terrazzoapp/terrazzo/commit/8f32d44792bba62194e674c9b60cfdeb366c96c7) Thanks [@michaelurban](https://github.com/michaelurban)! - feat: add typography shorthand, improve Sass plugin
 
+- Thanks [@drwpow](https://github.com/drwpow)! - Improve extended typography property CSS generation
+
 ## 0.10.3
 
 ### Patch Changes